+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
     <uses-permission android:name="android.permission.GET_ACCOUNTS" />
 
     <application
-        android:name="com.m.push.PushApplication"
+        android:name="com.m.services.push.PushApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -17,10 +17,9 @@
         android:theme="@style/AppTheme"
         android:networkSecurityConfig="@xml/network_security_config"
         tools:ignore="GoogleAppIndexingWarning">
-    <meta-data android:name="DEFAULT_MESSAGE_KEY" android:value="com.m.push.NotifyingHandler" />
+        <meta-data android:name="DEFAULT_MESSAGE_KEY" android:value="com.m.push.NotifyingHandler" />
 
-        <activity android:name="com.m.helper.LoginActivity">
-
+        <activity android:name="com.m.services.Auth.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
@@ -33,7 +32,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.m.dataSync.CreateTask"
+            android:name="com.m.services.dataSync.CreateTask"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="com.m.androidNativeApp.CreateTask"/>
@@ -41,7 +40,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.m.dataSync.SyncActivity"
+            android:name="com.m.services.dataSync.SyncActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="com.m.androidNativeApp.SyncActivity"/>
