+++ app/src/main/java/com/m/models/Item.java
@@ -1,4 +1,4 @@
-package com.m.helper;
+package com.m.models;
 
 public class Item {
