+++ app/src/main/java/com/m/services/Auth/AuthStateManager.java
@@ -1,4 +1,4 @@
-package com.m.services.appAuth;
+package com.m.services.Auth;
 
 
 import android.content.Context;
