+++ app/src/main/java/com/m/services/push/PushApplication.java
@@ -1,4 +1,4 @@
-package com.m.push;
+package com.m.services.push;
 
 import android.app.Application;
